refactor: Update MyChart widget to accept data as a parameter
This commit updates the MyChart widget in the chart.dart file to accept data as a parameter. The data is passed to the mainBarData and showingGroups methods, allowing for dynamic rendering of the chart based on the provided data. This change improves the flexibility and reusability of the MyChart widget.